SLING-11098 cleanup code smells and reduce duplication
